requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
# conda list -e > requirements.txt
# pip install -r requirements.txt
albumentations=1.2.0=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
astroid=2.9.0=py310hecd8cb5_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pyhd3eb1b0_1
black=19.10b0=py_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.4.26=hecd8cb5_0
certifi=2022.5.18.1=py310hecd8cb5_0
cffi=1.15.0=py310hc55c11b_1
cfgv=3.3.1=pypi_0
click=8.0.2=pypi_0
distlib=0.3.4=pypi_0
filelock=3.7.1=pypi_0
freetype=2.11.0=hd8bbffd_0
future=0.18.2=py310hecd8cb5_1
giflib=5.2.1=haf1e3a3_0
identify=2.5.1=pypi_0
imageio=2.19.3=pypi_0
intel-openmp=2021.4.0=hecd8cb5_3538
isort=5.9.3=pyhd3eb1b0_0
joblib=1.1.0=pypi_0
jpeg=9e=hca72f7f_0
lazy-object-proxy=1.6.0=py310hca72f7f_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=hdb42f99_1
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
lz4-c=1.9.3=h23ab428_1
mccabe=0.7.0=pyhd3eb1b0_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
mypy_extensions=0.4.3=py310hecd8cb5_1
ncurses=6.3=hca72f7f_2
networkx=2.8.4=pypi_0
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
nodeenv=1.6.0=pypi_0
numpy=1.22.3=py310hdcd3fac_0
numpy-base=1.22.3=py310hfd2de13_0
opencv-python-headless=4.6.0.66=pypi_0
openssl=1.1.1o=hca72f7f_0
packaging=21.3=pypi_0
pathspec=0.7.0=py_0
pillow=9.0.1=py310hde71d04_0
pip=21.2.4=py310hecd8cb5_0
platformdirs=2.4.0=pyhd3eb1b0_0
pre-commit=2.19.0=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pylint=2.12.2=py310hecd8cb5_1
pyparsing=3.0.9=pypi_0
python=3.10.4=hdfd78df_0
pytorch=1.10.2=cpu_py310h30e64cd_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.1.2=hca72f7f_1
regex=2022.3.15=py310hca72f7f_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.1=pypi_0
scipy=1.8.1=pypi_0
setuptools=61.2.0=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sqlite=3.38.3=h707629a_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.5.4=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd3eb1b0_0
tqdm=4.64.0=py310hecd8cb5_0
typed-ast=1.4.3=py310hca72f7f_1
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
virtualenv=20.14.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.13.3=py310hca72f7f_2
xz=5.2.5=hca72f7f_1
yapf=0.31.0=pyhd3eb1b0_0
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
